pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commonground-api-common"
version = "2.11.0"
description = "Commonground API tooling"
readme = "README.rst"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Maykin Media, VNG-Realisatie", email = "support@maykinmedia.nl" }
]
keywords = ["openapi", "swagger", "django"]
requires-python = ">=3.11"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "django>=5.2.0",
    "django-filter>=23.1,<=25.1",
    "django-solo",
    "djangorestframework>=3.15.0",
    "djangorestframework_camel_case>=1.2.0",
    "django-rest-framework-condition",
    "drf-nested-routers>=0.94.1",
    "drf-spectacular",
    "iso639-lang",
    "notifications-api-common>=0.3.1",
    "zgw-consumers>=1.0.0",
    "oyaml",
    "PyJWT>=2.10.1",
    "requests",
    "ape-pie",
    "sentry-sdk",
]

[project.urls]
Homepage = "https://github.com/maykinmedia/commonground-api-common"
Documentation = "https://commonground-api-common.readthedocs.io/en/latest/?badge=latest"
"Bug Tracker" = "https://github.com/maykinmedia/commonground-api-common/issues"
"Source Code" = "https://github.com/maykinmedia/commonground-api-common"

[project.optional-dependencies]
markdown_docs = [
    "django-markup<=1.3",
    "markdown",
]

djangorestframework_gis = [
    "djangorestframework-gis>=1.0",
]

drf_extra_fields = [
    "drf-extra-fields>=3.7.0",
]

tests = [
    "psycopg2",
    "pytest",
    "pytest-django",
    "pytest-dotenv",
    "pytest-factoryboy>=2.8.0",
    "tox",
    "ruff",
    "requests-mock",
    "freezegun",
    "zgw-consumers-oas",
    "djangorestframework-gis",
    "drf-extra-fields",
]
type-checking = [
    "pyright",
    "django-stubs",
    "django-stubs-ext",
]

oas = [
    "requests-mock",
]

testutils = [
    "zgw-consumers-oas",
]

setup-configuration = [
    "django-setup-configuration>=0.7.0",
]

notifications = [
    "notifications-api-common>=0.3.1",
]

coverage = [
    "pytest-cov",
]

docs = [
    "sphinx",
    "sphinx-rtd-theme",
]

release = [
    "bump-my-version",
]

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
include = ["vng_api_common*"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "testapp.settings"

[tool.coverage.run]
branch = true
source = ["vng_api_common"]
patch = ["subprocess"]
omit = [
    "*/migrations/*",
]

[tool.coverage.report]
skip_covered = true
exclude_also = [
    "if (typing\\.)?TYPE_CHECKING:",
    "@(typing\\.)?overload",
    "class .*\\(.*Protocol.*\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
    "\\bpass$",
]

[tool.bumpversion]
current_version = "2.11.0"
files = [
    {filename = "pyproject.toml"},
    {filename = "README.rst"},
    {filename = "docs/conf.py"},
    {filename = "vng_api_common/__init__.py"},
    {filename = "vng_api_common/locale/nl/LC_MESSAGES/django.po"},
]


[tool.ruff]
# Exclude folders and files
exclude = [
    "migrations",
    "static",
    "media",
]

# Same as Black
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I",
    # https://docs.astral.sh/ruff/rules/#error-e
    "E",
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F",
    # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "SIM",
    # https://docs.astral.sh/ruff/rules/#perflint-perf
    "PERF",
    # https://docs.astral.sh/ruff/rules/#refurb-furb
    "FURB",
    # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "B006",
    "B010",
    # https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "N",
    "TID251"
]

ignore = [
    # Checks for assertRaises and pytest.raises context managers that catch Exception or BaseException.
    "B017",
    # Checks for useless expressions.
    "B018",
    # Checks for raise statements in exception handlers that lack a from clause.
    "B904",
    # Whitespace before ':' (conflicts with Black)
    "E203",
    # Checks for lines that exceed the specified maximum character length.
    "E501",
    # Do not assign a lambda expression
    "E731",
    # Name may be undefined from '*' import
    "F405",
    # Checks for CamelCase imports that are aliased to lowercase names
    "N813",
    # Checks for CamelCase imports that are aliased to constant-style names.
    "N814",
    # Checks for class variable names that follow the mixedCase convention
    "N815",
    # Exception name should be named with Error suffix
    "N818",
    # Checks for instance methods that use a name other than self for their first argument.
    "N805",
    # Checks for the use of non-lowercase variable names in functions
    "N806",
    # Checks for functions names that do not follow the snake_case naming convention.
    "N802",
    # Checks for module names that do not follow the snake_case naming convention or are otherwise invalid.
    "N999",
    # Checks for for loops that can be replaced by a list comprehension.
    "PERF401",
    # Checks for `if` statements that can be replaced with bool
    "SIM103",
    # Check for if-else-blocks that can be replaced with a ternary operator
    "SIM108",
    #Check for environment variables that are not capitalized.
    "SIM112",
    # Checks for nested if statements that can be collapsed into a single if statement
    "SIM102",
    # Use a single `with` statement with multiple contexts instead of nested `with` statements
    "SIM117",
]


# Import sorting configuration (equivalent to isort)
[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = false
known-first-party = ["vng_api_common"]
lines-after-imports = -1
relative-imports-order = "furthest-to-closest"
sections = { django = ["django"] }
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.format]
line-ending = "lf"

[tool.setuptools.package-data]
open_api_framework = ["py.typed"]